Modifying primenumberchecker.py
The new code promotes reusability by using functions so that it can be incorporated in other modules. The code is made consice and cleaner. Also, the bug relating to giving '2' is handled. This code uses, square theorem hence in range function 'int(math.sqrt(num))' is included which is an advantage if the number is large.